test: add unexpected errors
