+++ tests/services/apiService.test.js
@@ -1,4 +1,4 @@
-import { fetchQuestCategoryData } from "@services/apiService";
+import { fetchQuestCategoryData, getQuestBoostClaimParams } from "@services/apiService";
 
 const API_URL = process.env.NEXT_PUBLIC_API_LINK;
 
@@ -40,3 +40,54 @@ describe("fetchQuestCategoryData function", () => {
     expect(result).toEqual(mockResponse);
   });
 });
+
+describe("getQuestBoostClaimParams function", () =>  {
+
+  beforeEach(() => {
+      fetch.mockClear();
+  }  
+  );
+  
+  it("should fetch and return data for a valid  id and address", async () => {
+         const mockData = ["string", "string"]; 
+         fetch.mockResolvedValueOnce({
+          json: () => Promise.resolve(mockData),
+        });
+  
+         const result = await getQuestBoostClaimParams ("string", "address");
+         expect(fetch).toHaveBeenCalledWith(
+          `${API_URL}/boost/get_claim_params?boost_id=string&addr=address`
+         );
+         expect(result).toEqual(mockData);
+  });
+  
+  it("should handle fetch errrors in unexpected format", async () => {
+      const mockResponse = "Boost with id 456 not found";
+  
+      fetch.mockResolvedValueOnce({
+       json: () => Promise.resolve(mockResponse),
+     });
+  
+     const result = await getQuestBoostClaimParams ("456", "566");
+     expect(fetch).toHaveBeenCalledWith(
+      `${API_URL}/boost/get_claim_params?boost_id=456&addr=566`
+     );
+     expect(result).toEqual(mockResponse);
+  })
+  
+  
+  it("should handle fetch with No Response ", async () => {
+      const mockResponse = "Failed to deserialize query string: invalid digit found in string";
+  
+      fetch.mockResolvedValueOnce({
+       json: () => Promise.resolve(mockResponse),
+     });
+  
+     const result = await getQuestBoostClaimParams ("string", "566");
+     expect(fetch).toHaveBeenCalledWith(
+      `${API_URL}/boost/get_claim_params?boost_id=string&addr=566`
+     );
+     expect(result).toEqual(mockResponse);
+  })
+  
+  })
